[FR-810]
- Creating sub module for spring
- Some refactoring go allow user to configure his own context resolution
