--- authenticator/src/main/java/com/frontegg/sdk/middleware/authenticator/FronteggAuthentication.java
@@ -1,9 +1,0 @@
-package com.frontegg.sdk.middleware.authenticator;
-
-import java.time.LocalDateTime;
-
-public interface FronteggAuthentication {
-    LocalDateTime getAccessTokenExpiry();
-    String getAccessToken();
-
-}
