+++ config/pom.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
+        <groupId>com.frontegg.sdk</groupId>
         <artifactId>java-sdk</artifactId>
-        <groupId>com.frontegg</groupId>
         <version>0.0.2-SNAPSHOT</version>
     </parent>
 	<modelVersion>4.0.0</modelVersion>
@@ -13,7 +13,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.frontegg</groupId>
+            <groupId>com.frontegg.sdk</groupId>
             <artifactId>common</artifactId>
             <version>${project.version}</version>
         </dependency>
