--- frontegg-springboot-sample/src/main/java/com/frontegg/samples/Application.java
@@ -1,15 +1,0 @@
-package com.frontegg.samples;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application
-{
-
-	public static void main(String[] args)
-	{
-		SpringApplication.run(Application.class, args);
-	}
-
-}
