--- frontegg-springboot-sample/src/main/java/com/frontegg/samples/SamlRequest.java
@@ -1,17 +1,0 @@
-package com.frontegg.samples;
-
-public class SamlRequest
-{
-	private String payload;
-
-	public SamlRequest()
-	{
-	}
-
-	public String getPayload()
-	{
-		return payload;
-	}
-
-
-}
