--- frontegg-springboot-sample/src/main/resources/application.yml
@@ -1,3 +1,0 @@
-frontegg:
-  clientId: "<YOUR CLIENT ID>"
-  apiKey: "<YOUR API KEY>"
