--- middleware-spring/src/main/java/com/frontegg/sdk/middleware/spring/config/DefaultFronteggConfigProviderChain.java
@@ -1,20 +1,0 @@
-package com.frontegg.sdk.middleware.spring.config;
-
-import com.frontegg.sdk.config.DefaultConfigProvider;
-import com.frontegg.sdk.config.EnvironmentVariableConfigProvider;
-import com.frontegg.sdk.config.SystemPropertiesConfigProvider;
-
-public class DefaultFronteggConfigProviderChain extends FronteggConfigProviderChain {
-
-    private static final DefaultFronteggConfigProviderChain INSTANCE = new DefaultFronteggConfigProviderChain();
-
-    public DefaultFronteggConfigProviderChain() {
-        super(
-           new EnvironmentVariableConfigProvider(),
-           new SystemPropertiesConfigProvider(),
-           new DefaultConfigProvider()
-        );
-    }
-
-    public static DefaultFronteggConfigProviderChain getInstance() {return INSTANCE;}
-}
