--- middleware-springboot-autoconfigure/pom.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>java-sdk</artifactId>
-        <groupId>com.frontegg</groupId>
-        <version>0.0.2-SNAPSHOT</version>
-    </parent>
-	<modelVersion>4.0.0</modelVersion>
-	
-	<packaging>jar</packaging>
-
-    <artifactId>middleware-springboot-autoconfigure</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-autoconfigure</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.frontegg</groupId>
-            <artifactId>middleware-spring</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-</project>
