--- middleware-springboot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.frontegg.middleware.spring.config.FronteggAutoConfiguration
