--- middleware/src/main/java/com/frontegg/sdk/middleware/authentication/FronteggAuthenticationService.java
@@ -1,16 +1,0 @@
-package com.frontegg.sdk.middleware.authentication;
-
-import javax.servlet.http.HttpServletRequest;
-
-public interface FronteggAuthenticationService {
-
-    /**
-     * Validate request is authorized
-     */
-    void withAuthentication(HttpServletRequest request);
-
-    /**
-     * Authenticate Frontegg application If not authenticated yet.
-     */
-    void authenticateFronteggApplicationIfNeeded();
-}
