+++ middleware/src/main/java/com/frontegg/sdk/middleware/context/FronteggContextResolver.java
@@ -1,0 +1,8 @@
+package com.frontegg.sdk.middleware.context;
+
+import javax.servlet.http.HttpServletRequest;
+
+public interface FronteggContextResolver
+{
+	void resolveContext(HttpServletRequest request);
+}
