--- middleware/src/main/java/com/frontegg/sdk/middleware/identity/FronteggIdentityService.java
@@ -1,6 +1,0 @@
-package com.frontegg.sdk.middleware.identity;
-
-public interface FronteggIdentityService {
-
-    void verifyToken(String token);
-}
