+++ spring/spring-frontegg-identity-context-resolver/pom.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <groupId>com.frontegg.sdk.spring</groupId>
+        <artifactId>spring</artifactId>
+        <version>0.0.2-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>spring-frontegg-identity-context-resolver</artifactId>
+
+    <properties>
+        <auth0.jwt.version>3.2.0</auth0.jwt.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.frontegg.sdk</groupId>
+            <artifactId>middleware</artifactId>
+            <version>${project.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.auth0</groupId>
+            <artifactId>java-jwt</artifactId>
+            <version>${auth0.jwt.version}</version>
+        </dependency>
+
+    </dependencies>
+
+</project>
